requirements.txt
alembic==1.7.4
anyio==3.3.4
apache-airflow==2.2.2
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
appnope==0.1.2
argcomplete==1.12.3
async-generator==1.10
attrs==20.3.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
cattrs==1.5.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
chromedriver-binary==91.0.4472.19.0
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.1.1
distlib==0.3.2
distro==1.5.0
dnspython==2.1.0
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
easyocr==1.4.1
email-validator==1.1.3
entrypoints==0.3
filelock==3.0.12
Flask==1.1.2
Flask-AppBuilder==3.3.4
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-MySQLdb==0.2.0
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.18
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
html5==0.0.9
httpcore==0.13.7
httpx==0.20.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.2
inflection==0.5.1
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
iso8601==1.0.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
kaggle==1.5.12
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.6.3
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.14.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.20.3
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
opencv-python-headless==4.5.4.60
outcome==1.1.0
packaging==21.0
pandas==1.2.4
paramiko==2.7.2
parso==0.8.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
platformdirs==2.3.0
prison==0.2.1
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
psycopg2==2.9.1
psycopg2-binary==2.9.2
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.21
Pygments==2.10.0
PyJWT==1.7.1
PyMySQL==1.0.2
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-bidi==0.4.2
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.17.0
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyWavelets==1.2.0
PyYAML==5.4.1
pyzmq==22.2.1
requests==2.26.0
rfc3986==1.5.0
rich==10.13.0
scikit-image==0.19.0
scipy==1.7.3
selenium==4.0.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sox==1.4.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.9
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
texttable==1.6.3
tifffile==2021.11.2
toml==0.10.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.1.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
unicodecsv==0.14.1
urllib3==1.26.7
virtualenv==20.8.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
wsproto==1.0.0
WTForms==2.3.3
zipp==3.6.0
